Documentation and Shannon Entropy Correction
- Added Documentation to unknown variables
- Corrected Shannon Entropy function to look at number of tiles instead of calculation from probability.
