--- map.kv
@@ -1,10 +1,0 @@
-<Map>:
-    layout:_layout
-    size_hint: None, None
-    do_rotation: False
-    auto_bring_to_front: False
-
-    FloatLayout:
-        id: _layout
-        size_hint: (None, None)
-        pos: (0, 0)
