+++ ui/map.kv
@@ -1,0 +1,10 @@
+<Map>:
+    layout:_layout
+    size_hint: None, None
+    do_rotation: False
+    auto_bring_to_front: False
+
+    FloatLayout:
+        id: _layout
+        size_hint: (None, None)
+        pos: (0, 0)
